+++ NEWS.md
@@ -1,3 +1,8 @@
+# lz4lite 0.1.1
+
+* Added support for vectors of complex numbers
+* Minor refactor of C code to use `DATAPTR`
+
 # lz4lite 0.1.0
 
 * Initial release